+++ inst/include/cpp11/doubles.hpp
@@ -131,6 +131,27 @@ typedef r_vector<double> doubles;
 
 }  // namespace writable
 
+typedef r_vector<int> integers;
+
+inline doubles as_double(sexp x) {
+  if (TYPEOF(x) == REALSXP) {
+    return as_cpp<doubles>(x);
+  }
+
+  else if (TYPEOF(x) == INTSXP) {
+    integers xn = as_cpp<integers>(x);
+    size_t len = xn.size();
+    writable::doubles ret;
+    for (size_t i = 0; i < len; ++i) {
+      ret.push_back(static_cast<double>(xn[i]));
+    }
+    return ret;
+  }
+
+  // else
+  stop("Expected a numeric vector for");
+}
+
 template <>
 inline double na() {
   return NA_REAL;